remove types
